ImageStatistics/requirements.txt
autopep8==1.5.5
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.1
six==1.15.0
toml==0.10.2
